+++ src/components/common/ErrorText.tsx
