+++ src/components/common/FormCamera.tsx
@@ -1,7 +1,7 @@
 import React, { Dispatch } from 'react';
 import { Image, Text, View } from 'react-native';
 
-import { Button, CameraButton, ImageButton } from './Buttons';
+import { Button, CameraButton, ImageButton } from './buttons';
 
 import { CatalogImageHandler } from '@/image_handlers/CatalogImageHandler';
 import { buttonStyles, containerStyles, textStyles } from '@/styles';
