+++ src/components/common/buttons/index.js
@@ -1,0 +1,2 @@
+export * from './Button';
+export * from './CameraButton';
