+++ src/components/index.js
@@ -1,11 +1,4 @@
-export * from './ui/Buttons';
-export { DateTimeInput } from './ui/DateTimeInput';
-export { ErrorText } from './ui/ErrorText';
-export { LoadingIndicator } from './ui/LoadingIndicator';
-export { Switch } from './ui/Switch';
-export * from './ui/TextInput';
-export { SnackbarMessage } from './ui/SnackBarMessage';
-export { FormCamera } from './ui/FormCamera';
+export * from './common';
 
 export { AdminView } from './AdminView';
 export { Errorbar } from './ErrorBar';
