+++ src/components/items/UserItem.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Text, View } from 'react-native';
 
-import DatabaseService from '../../services/DatabaseService';
-import { Button } from '../ui/Buttons';
+import { Button } from '../common';
 import { useRouter } from 'expo-router';
 
 import { useAuth } from '@/providers';
+import DatabaseService from '@/services/DatabaseService';
 import { buttonStyles, containerStyles, textStyles } from '@/styles';
 import { User } from '@/types';
